Testsuite fixups

+++ gcc/testsuite/g++.dg/torture/pr33887-2.C
@@ -1,17 +1,19 @@
-/* { dg-do run } */
+/* { dg-do run { target { stdint_types } } } */
+
+#include <stdint.h>
 
 extern "C" void abort() __attribute__ ((noreturn));
 
 struct s
 {
   unsigned long long f1 : 40;
-  unsigned int f2 : 24;
+  uint_least32_t f2 : 24;
 } sv;
 
 int main()
 {
-  int f2;
-  sv.f2 = (1 << 24) - 1;
+  int_least32_t f2;
+  sv.f2 = ((uint_least32_t)1 << 24) - 1;
   __asm__ volatile ("" : : : "memory");
   ++sv.f2;
   f2 = sv.f2;
