+++ gcc/testsuite/gcc.dg/pr64536.c
@@ -1,11 +1,13 @@
 /* PR rtl-optimization/64536 */
-/* { dg-do link } */
+/* { dg-do link { target { stdint_types } } } */
 /* { dg-options "-O2" } */
 /* { dg-additional-options "-fPIC" { target fpic } } */
 
-struct S { long q; } *h;
-long a, b, g, j, k, *c, *d, *e, *f, *i;
-long *baz (void)
+#include <stdint.h>
+
+struct S { intptr_t q; } *h;
+intptr_t a, b, g, j, k, *c, *d, *e, *f, *i;
+intptr_t *baz (void)
 {
   asm volatile ("" : : : "memory");
   return e;
@@ -39,8 +41,8 @@ bar (int x)
 	      h->q = *f;
 	    }
 	  else
-	    i = (long *) (h->q = *f);
-	  *c++ = (long) f;
+	    i = (intptr_t *) (h->q = *f);
+	  *c++ = (intptr_t) f;
 	  e += 6;
 	}
       else
@@ -53,8 +55,8 @@ bar (int x)
 	      h->q = *f;
 	    }
 	  else
-	    i = (long *) (h->q = *f);
-	  *c++ = (long) f;
+	    i = (intptr_t *) (h->q = *f);
+	  *c++ = (intptr_t) f;
 	  e += 6;
 	}
     }
